pyproject.toml
[project]
name = "obss"
version = "0.1.0"
description = "An Obsidian semantic searcher."
authors = [
    {name = "Sean Chou", email = "sean@choufam.com"},
]
dependencies = [
    "about-time>=4.2.1",
    "aiohttp>=3.12.14",
    "aiosignal>=1.3.1",
    "alive-progress>=3.1.4",
    "async-timeout>=4.0.3",
    "attrs>=23.1.0",
    "certifi>=2023.7.22",
    "charset-normalizer>=3.2.0",
    "contourpy>=1.1.0",
    "cycler>=0.11.0",
    "dill>=0.3.7",
    "fonttools>=4.42.1",
    "frozenlist>=1.4.0",
    "grapheme>=0.6.0",
    "idna>=3.4",
    "joblib>=1.3.2",
    "kiwisolver>=1.4.5",
    "matplotlib>=3.10.0",
    "multidict>=6.0.4",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "packaging>=23.1",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "patsy>=0.5.3",
    "pillow>=10.0.0",
    "plotly>=6.2.0",
    "pyfunctional>=1.4.3",
    "pyparsing>=3.0.9",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "regex>=2023.8.8",
    "requests>=2.31.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.11.2",
    "seaborn>=0.12.2",
    "six>=1.16.0",
    "statsmodels>=0.14.0",
    "tabulate>=0.9.0",
    "tenacity>=8.2.3",
    "threadpoolctl>=3.2.0",
    "tiktoken>=0.4.0",
    "tqdm>=4.66.1",
    "tzdata>=2023.3",
    "urllib3>=2.5.0",
    "yarl>=1.9.2",
]
readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    # Add development dependencies here, e.g.:
    # "pytest>=8.0.0",
    # "black>=24.0.0",
    # "ruff>=0.1.0",
]